NFC: trf7970a: Set 'Modulator and SYS_CLK Control' after 'ISO Control'
Writing to the 'ISO Control' register may cause the
contents of the 'Modulator and SYS_CLK Control' register
to change so be sure to write to 'Modulator and SYS_CLK
Control' after writing to 'ISO Control'.

Note that writing to the 'Modulator and SYS_CLK Control'
register shouldn't be necessary at all according to the
trf790a manual but testing shows that it is necessary.

Signed-off-by: Mark A. Greer <mgreer@animalcreek.com>