NFC: trf7970a: Add pm_runtime support
Add pm_runtime support by moving the code
that enables the trf7970a to the pm_runtime
hook routines.  The pm_runtime 'autosuspend'
feature is used so that the device isn't
disabled until at least 30 seconds have
passed since trf7970a_switch_rf_off() was
last called.

The result is that when trf7970a_switch_rf_on()
is called, the device will be enabled and
initialized (if it isn't already).  When
trf7970a_switch_rf_off() is called, it will
turn off the RF immediately but leave the
device enabled for at least 30 seconds.
If 30 seconds have passed and the pm_runtime
facility decides to suspend the driver, the
device will be disabled then.

Signed-off-by: Mark A. Greer <mgreer@animalcreek.com>