NFC: trf7970a: Allow different Modulator and SYS_CLK Control register values
Currently the driver writes the same value to the
'Modulator and SYS_CLK Control' register no matter
what RF technology is being used.  That works for
now but new RF technologies (e.g., ISO/IEC 14443-B)
will require different values to be written to that
register.  To support this, add a member to the
trf7970a structure which is set by the RF technology
handling code and used by the framing code when it
writes to that register.

Signed-off-by: Mark A. Greer <mgreer@animalcreek.com>