NFC: trf7970a: Only reset FIFO when no more TX bytes left
Commit d0a0397c45240cb7e0b9e2cdd192c412b5df5a19
(NFC: trf7970a: Reset FIFO when 'End of TX' Interrupt Occurs)
unconditionally resets the FIFO when the trf7970a generates
an interrupt with the TX bit set.  However, an interrupt with
that bit set can be generated when the TX is not complete and
the FIFO just went past its low watermark.  In this case, the
FIFO should not be cleared (as it would destroy the remainder
of the TX frame) and instead just ignore the interrupt.  Once
the TX is complete, the trf7970a will generate another interrupt
with the TX bit set but not the FIFO bit so the FIFO can be
cleared then.

Signed-off-by: Mark A. Greer <mgreer@animalcreek.com>