HACK: Extend timeouts & Check FIFO status on IRQ
Extend timeouts to larger value since we're timing out too soon.

When timing out and haven't sent all TX data yet, read the
FIFO_STATUS register and if there are still bytes in it,
restart the timer.

When timing out and waiting for RX data, read the FIFO_STATUS
register and if there are bytes in it, reset the timer and
if there aren't, send the RX data upstream.

Signed-off-by: Mark A. Greer <mgreer@animalcreek.com>