HACK: Read IRQ_STATUS before sending new pkt, extra bytes in FIFO
Sometimes the trf7970a asserts the IRQ line for no
apparent reason which can confuse the driver when
its sending a new command.  This will sometimes
cause a 5.6.2.1 error to be emitted from the
digital layer.  To work around this error,
read the IRQ_STATUS register just before sending
a new command to clear the interrupt.

Also, don't count the bytes that aren't going to the peer device
in the count of FIFO bytes.

Signed-off-by: Mark A. Greer <mgreer@animalcreek.com>